REF (string): de-duplicate ArrowStringArray methods (#59555)
